+++ packages/cubejs-testing/birdbox-fixtures/postgresql/scripts/load.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
+exec 2>&1
+set -x
+set -e
 
 psql -U test -d test -f /data/*.sql