+++ packages/cubejs-testing/test/birdbox-postgresql-cubestore.test.ts
@@ -3,5 +3,5 @@ import { startBirdBoxFromContainer } from '../src';
 
 createBirdBoxTestCase('postgresql-cubestore', () => startBirdBoxFromContainer({
   name: 'postgresql-cubestore',
-  loadScript: 'load.sh',
+  loadScript: 'postgres-load-events.sh',
 }));